refactor(UserInfoDTO): 파라미터 타입 변경
